// src/models/Team.ts
export type Team = {
  id: string;
  name: string;
  shortName: string;
  city: string;
  state: string;
  captain: string;
  homeCourse: string;
  logo: string;
  primaryColor: string;
  secondaryColor: string;
  website: string;
  facebook: string;
  description: string;
  active: boolean;
  createdAt: string;
  updatedAt: string;
};